backend/requirements.txt
fastapi
uvicorn

sqlalchemy
psycopg2-binary
python-dotenv

bcrypt
pyjwt
pypdf
chromadb

google-generativeai

langchain
langchain-community
langgraph

pymupdf
python-docx

typing_extensions
pydantic
pydantic-settings
